requirements.txt
python-dotenv
requests
langchain
langchain-openai

fastapi
uvicorn
llama-index
pydantic
google-generativeai